Fix PlanNode.Plan.Actual Rows int64 unmarshal error on fractional averages
PostgreSQL computes "Actual Rows" as total_rows / nloops and always
formats it with 2 decimal digits in JSON EXPLAIN output (ExplainNode()
in explain.c), for every node type. When Actual Loops > 1 and the
division doesn't come out even (e.g. 3079.00, or 242.31 for
high-loop-count inner scans), the value is a genuine non-integer float.
pgplan modeled ActualRows as int64, so json.Unmarshal rejected any plan
with loops > 1, which is common for nested loop inner sides — this
crashed `pgplan compare`/`analyze` on real-world plans.

Change ActualRows (and, defensively, RowsRemovedByFilter/
RowsRemovedByJoinFilter, which share the same instrumentation path)
to float64, and propagate the type through the analyzer rules,
comparator diff/result structs, and text output formatting. Other
int64 fields (Plan Rows, Actual Loops, buffer/block counts, hash/sort
stats) were audited against PostgreSQL's source and confirmed to
always be emitted as whole numbers, so they're left unchanged.

Adds regression tests covering fractional values mixed with plain
integers in the same plan, and confirms the fix holds across multiple
node types (Seq Scan, Index Scan, Index Only Scan, Bitmap Heap Scan,
Hash Join, Sort, Aggregate) since the underlying Postgres code path is
shared by all node types.

+++ internal/comparator/result.go
@@ -66,9 +66,9 @@ type NodeDelta struct {
 	TimePct   float64
 	TimeDir   Direction
 
-	OldRows   int64
-	NewRows   int64
-	RowsDelta int64
+	OldRows   float64
+	NewRows   float64
+	RowsDelta float64
 	RowsPct   float64
 	RowsDir   Direction
 
@@ -77,8 +77,8 @@ type NodeDelta struct {
 	NewLoops int64
 
 	// Filter effectiveness
-	OldRowsRemovedByFilter int64
-	NewRowsRemovedByFilter int64
+	OldRowsRemovedByFilter float64
+	NewRowsRemovedByFilter float64
 
 	// Parallel
 	OldWorkersLaunched int
